feat: parse incoming requests and call service methods to process the logic for user routes

+++ repository/user_repository.go
@@ -8,7 +8,7 @@ import (
 )
 
 type UserRepositoryI interface {
-	FindByID(id int) (model.User, error)
+	FindByID(id uint64) (model.User, error)
 	Create(user model.User, tx ...*gorm.DB) (uint64, error)
 	Save(user *model.User) error
 }
@@ -21,7 +21,7 @@ func NewUserRepository(db *db.Database) UserRepositoryI {
 	return &UserRepository{storage: db}
 }
 
-func (repo *UserRepository) FindByID(id int) (model.User, error) {
+func (repo *UserRepository) FindByID(id uint64) (model.User, error) {
 	var user model.User
 	err := repo.storage.Exec(db.IRead).Where("id = ?", id).First(&user).Error
 	return user, err
